# pandora_console/api/v1/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nueva API de Pandora FMS",
    "description": "Nueva Api de pandora FMS",
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "Nombre del contacto",
      "url": "https://www.example.com/support",
      "email": "contacto@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "PandoraFMS API Server"
    }
  ],
  "paths": {
...
    "/user/{idUser}/profile/{idProfile}": {
      "get": {
        "tags": ["Users"],
        "summary": "show data field user profile",
        "operationId": "6aacb61062bd1ec04b91907273e2dd8f",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdUser"
          },
          {
            "$ref": "#/components/parameters/parameterIdProfile"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseUserProfile"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create user profile",
        "operationId": "9fe3300b6a0e031989b085cdba8def4d",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdUser"
          },
          {
            "$ref": "#/components/parameters/parameterIdProfile"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/requestBodyUserProfile"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseUserProfile"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Deletes user profile.",
        "operationId": "4fc44af32440fc1b9feff41f92c63972",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdUser"
          },
          {
            "$ref": "#/components/parameters/parameterIdProfile"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/successfullyDeleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/user/{idUser}/profiles": {
      "post": {
        "tags": ["Users"],
        "summary": "List user profiles",
        "operationId": "3952d52c24c4fd099e9c0d15c7104068",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdUser"
          },
          {
            "$ref": "#/components/parameters/parameterPage"
          },
          {
            "$ref": "#/components/parameters/parameterSizePage"
          },
          {
            "$ref": "#/components/parameters/parameterSortField"
          },
          {
            "$ref": "#/components/parameters/parameterSortDirection"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/requestBodyUserProfile"
        },
        "responses": {
          "200": {
            "description": "List data profiles user object",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "paginationData": {
                      "$ref": "#/components/schemas/paginationData"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserProfile"
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "paginationData": {
        "description": "Info pagination data",
        "properties": {
          "totalPages": {
            "description": "Number of pages",
            "type": "integer",
            "readOnly": true,
            "nullable": true
          },
          "sizePage": {
            "description": "Items per page",
            "type": "integer",
            "readOnly": true,
            "nullable": true
          },
          "totalRegisters": {
            "description": "Number of items",
            "type": "integer",
            "readOnly": true,
            "nullable": true
          },
          "totalRegistersPage": {
            "description": "Number of items this page",
            "type": "integer",
            "readOnly": true,
            "nullable": true
          },
          "currentPage": {
            "description": "Number of current page",
            "type": "integer",
            "readOnly": true,
            "nullable": true
          }
        },
        "type": "object"
      },
...
      "UserProfile": {
        "properties": {
          "idUserProfile": {
            "description": "Id user profile",
            "type": "integer",
            "readOnly": true,
            "nullable": false
          },
          "idUser": {
            "description": "Id User",
            "type": "string",
            "readOnly": true,
            "nullable": false
          },
          "idProfile": {
            "description": "Id Profile",
            "type": "integer",
            "readOnly": true,
            "nullable": false
          },
          "idGroup": {
            "description": "Id Group",
            "type": "integer",
            "nullable": false
          },
          "isNoHierarchy": {
            "description": "No hierarchy",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "assignedBy": {
            "description": "Create user profile by",
            "type": "string",
            "default": null,
            "readOnly": true,
            "nullable": false
          },
          "idPolicy": {
            "description": "Id Policy",
            "type": "integer",
            "nullable": false
          },
          "tags": {
            "description": "Tags to which a user belongs",
            "type": "array",
            "items": {
              "properties": {
                "idTags": {
                  "description": "Tags id",
                  "type": "integer",
                  "default": null,
                  "nullable": true
                }
              },
              "type": "object"
            },
            "default": null,
            "nullable": true
          }
        },
        "type": "object"
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "successfullyDeleted": {
        "description": "Successfully deleted",
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "result": {
                  "type": "string",
                  "default": "Successfully deleted"
                }
              },
              "type": "object"
            }
          }
        }
      },
...
      "ResponseUserProfile": {
        "description": "User Profile type object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UserProfile"
            }
          }
        }
      }
    },
    "parameters": {
      "parameterIdProfile": {
        "name": "idProfile",
        "in": "path",
        "description": "Profile id",
        "required": true,
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "parameterPage": {
        "name": "page",
        "in": "query",
        "description": "page",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 0
        }
      },
      "parameterSizePage": {
        "name": "sizePage",
        "in": "query",
        "description": "Size page",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 0
        }
      },
      "parameterSortField": {
        "name": "sortField",
        "in": "query",
        "description": "sort field",
        "required": false,
        "schema": {
          "type": "string",
          "default": ""
        }
      },
      "parameterSortDirection": {
        "name": "sortDirection",
        "in": "query",
        "description": "sort direction",
        "required": false,
        "schema": {
          "type": "string",
          "default": "",
          "enum": ["ascending", "descending"]
        }
      },
      "parameterIdUser": {
        "name": "idUser",
        "in": "path",
        "description": "User id",
        "required": true,
        "schema": {
          "type": "string",
          "default": "admin"
        }
      },
...
    },
    "requestBodies": {
...
      "requestBodyUserProfile": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UserProfile"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "in": "header",
        "bearerFormat": "pandoraBearer",
        "scheme": "bearer"
      }
    }
  },
...
}